vision/scripts/requirements.txt
azure-cognitiveservices-vision-customvision==3.1.0
azure-cognitiveservices-vision-face==0.4.1
azure-common==1.1.26
msrest==0.6.19
Pillow==7.0.0
PySimpleGUI==4.32.1
screen-resolution-extra==0.0.0
screeninfo==0.6.6